eclipse/rdf4j#1453 TripleStore returns count-per-context now

+++ nativerdf/src/main/java/org/eclipse/rdf4j/sail/nativerdf/ContextStore.java
@@ -109,7 +109,7 @@ public class ContextStore implements Iterable<Resource> {
 	 * @param context the context identifier.
 	 * @param amount  the number by which to decrease the size
 	 */
-	public void decrementBy(Resource context, int amount) {
+	public void decrementBy(Resource context, long amount) {
 		contextInfoMap.computeIfPresent(context, (c, size) -> size <= amount ? null : size - amount);
 		contentsChanged = true;
 	}
